리스트형 article 구조 수정
